cloudrun_functions/scrape_news/requirements.txt
functions-framework==3.*
google-cloud-secret-manager
google-cloud-storage
duckdb==1.1.0
pandas
yfinance
requests
bs4
